Extract parser related code from `lib`
